+++ modules/web/web-app/.eslintrc.js
@@ -3,6 +3,7 @@ module.exports = {
   // load the config from `eslint-config-meteroid`
   extends: ['meteroid'],
   rules: {
+    'import/no-cycle': 'error',
     '@typescript-eslint/no-unused-vars': [
       'error',
       {
