+++ modules/web/web-app/App.tsx
@@ -4,7 +4,11 @@ import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { RouterProvider } from 'react-router-dom'
 import { Toaster } from 'sonner'
 
-import { authInterceptor, errorInterceptor, loggingInterceptor } from '@/lib/connectrpc'
+import {
+  authInterceptor,
+  errorInterceptor,
+  loggingInterceptor,
+} from '@/lib/connectrpc-interceptors'
 import { env } from '@/lib/env'
 
 import router from './router/router'
