+++ modules/web/web-app/features/billing/plans/pricecomponents/components/shared.ts
@@ -1,0 +1,6 @@
+import { FormPriceComponent } from '@/lib/schemas/plans'
+
+export interface FeeFormProps {
+  cancel: () => void
+  onSubmit: (data: FormPriceComponent['fee']['data']) => void
+}
