+++ modules/web/web-app/package.json
@@ -83,6 +83,7 @@
     "tailwindcss-radix": "^2.8.0",
     "typescript": "^5.3.3",
     "vite": "^5.0.13",
+    "vite-plugin-circular-dependency": "^0.5.0",
     "vite-plugin-svgr": "^4.2.0",
     "vite-tsconfig-paths": "^4.3.1"
   },
