+++ modules/web/web-app/vite.config.ts
@@ -1,6 +1,7 @@
 import basicSsl from '@vitejs/plugin-basic-ssl'
 import react from '@vitejs/plugin-react'
 import { UserConfigExport, defineConfig } from 'vite'
+import circleDependency from 'vite-plugin-circular-dependency'
 import svgr from 'vite-plugin-svgr'
 import tsconfigPaths from 'vite-tsconfig-paths'
 
@@ -22,7 +23,7 @@ export default defineConfig(({ mode }) => {
       },
     },
     envDir: '../../../',
-    plugins: [react(), tsconfigPaths(), svgr()],
+    plugins: [react(), tsconfigPaths(), svgr(), circleDependency()],
   }
   return localSsl
     ? {
